POSIXCore: adjust Linux path for the signal handler routine
When building against the GNU C library with modern Swift compilers,
`_POSIX_SOURCE` is defined which will enable the IEEE 1003.1 compliance
which in turn will rename the `sa_handler` member and provide a
convenience macro. Unfortunately, the macro is not imported into Swift
today and that requires that we manually expand it inline here.

Thanks to @zaneenders for reporting this!